data-gen/requirements.txt
confluent-kafka==2.6.0
psycopg2-binary==2.9.9
redis==5.1.1
Faker==28.4.1
requests==2.32.3
